smart_cat_water_bowl/board/requirements.txt
google-cloud-firestore>=2.10.0
RPi.GPIO==0.7.0
